+++ R/map.R
@@ -94,12 +94,12 @@ setMethod("plot3D", "MSmap",
               if (rgl) {
                   if (!require("rgl"))
                       stop("The 'rgl' package needed. Install it with 'install.packages(\"rgl\")'.")
-                  rgl::plot3d(rev(dd$mz), dd$rt, dd$intensity, , type = "h",
+                  rgl::plot3d(dd$mz, dd$rt, dd$intensity, , type = "h",
                          xlab = "M/Z", ylab = "Retention time", zlab = "")
               } else {                  
                   ms <- NULL ## get rid of 'no visible global function definition' note
                   par.set <- list(box.3d = list(lwd=.2))
-                  cloud(intensity ~ rev(mz) + rt , data = dd,
+                  cloud(intensity ~ mz + rt , data = dd,
                         type="h",
                         scales= list(
                             arrows=FALSE,