+++ NEWS.md
@@ -3,6 +3,8 @@
 ## Changes in version 2.5.3
 - Add filterPrecursorScan for `MSnExp` and `OnDiskMSnExp`;
   closes issue #282 and PR #287 <2017-12-16 Sat>.
+- MSnSet to/from SummarizedExperiment coercion (contributed by Arne
+  Smits in PR [#284](https://github.com/lgatto/MSnbase/issues/284)) <2017-12-17 Sun>
 
 ## Changes in version 2.5.2
 - Use automatic backend detection (based on file name and file