+++ NEWS.md
@@ -1,5 +1,16 @@
 # MSnbase 2.5
 
+## Changes in version 2.5.4
+- Add featureData slot to Chromatograms class and add mz, precursorMz and 
+  productMz methods for Chromatograms (issue
+  [#289](https://github.com/lgatto/MSnbase/issues/289)) <2017-12-18 Mon>.
+
+## Changes in version 2.5.3
+- MSnSet to/from SummarizedExperiment coercion (contributed by Arne
+  Smits in PR [#284](https://github.com/lgatto/MSnbase/issues/284)) <2017-12-17 Sun>
+- Fix inverted M/Z axis in plot3D,MSmap (reported by Sylvain
+  Dechaumet, see issue [#292](https://github.com/lgatto/MSnbase/issues/292)) <2017-12-19 Tue>
+
 ## Changes in version 2.5.2
 - Use automatic backend detection (based on file name and file
   content) that was introduced in mzR version 2.13.1 (issue